Continued implementing Actions.
Got rid of "states" as separate entities and replaced it with "opening" and "closing" an Action. "open" the Action at the beginning of a method being called (with timestamp = NULL), and "close" it (add a timestamp) at the end of the method.
Note that some submethods may try to "close()" the Action, but that should not be allowed (via action.closeable = False).

But now how do I model the relationships between objects?! I mostly figured out objects & their attributes, but what about between objects?!